Add sync committee rewards to beacon chain trait
